# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NEXUS Super Social Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List / search users
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: Feed posts
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [photo, video, text, all]
            default: all
        - in: query
          name: userId
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
...
  /reels:
    get:
      operationId: listReels
      tags: [reels]
      summary: Get reels feed
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 10 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
        - in: query
          name: userId
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reel"
    post:
      operationId: createReel
      tags: [reels]
      summary: Upload a reel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReelInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reel"
...
  /stories:
    get:
      operationId: listStories
      tags: [stories]
      summary: Get active stories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Story"
    post:
      operationId: createStory
      tags: [stories]
      summary: Post a story
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
...
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List communities/groups
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a group/community
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: Get user notifications
      parameters:
        - in: query
          name: unread
          schema: { type: boolean }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin list all users with filters
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [active, suspended, all]
            default: all
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
...
  /creator/plans:
    post:
      operationId: createCreatorPlan
      tags: [creator]
      summary: Create a subscription plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatorPlanInput"
      responses:
        "201":
          description: Plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorPlan"
...
  /search:
    get:
      operationId: searchAll
      tags: [search]
      summary: Search users, posts, reels, and products
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string, enum: [all, users, posts, reels, products] }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, displayName, email, createdAt]
      properties:
        id: { type: integer }
        username: { type: string }
        displayName: { type: string }
        email: { type: string }
        bio: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        coverUrl: { type: ["string", "null"] }
        followersCount: { type: integer }
        followingCount: { type: integer }
        postsCount: { type: integer }
        isVerified: { type: boolean }
        isFollowing: { type: boolean }
        status: { type: string, enum: [active, suspended] }
        createdAt: { type: string }
    UserInput:
      type: object
      required: [username, displayName, email]
      properties:
        username: { type: string }
        displayName: { type: string }
        email: { type: string }
        bio: { type: string }
        avatarUrl: { type: string }
...
    Post:
      type: object
      required: [id, author, content, type, likesCount, commentsCount, createdAt]
      properties:
        id: { type: integer }
        author:
          $ref: "#/components/schemas/User"
        content: { type: string }
        type: { type: string, enum: [photo, video, text] }
        mediaUrl: { type: ["string", "null"] }
        thumbnailUrl: { type: ["string", "null"] }
        tags: { type: array, items: { type: string } }
        likesCount: { type: integer }
        commentsCount: { type: integer }
        sharesCount: { type: integer }
        isLiked: { type: boolean }
        isFlagged: { type: boolean }
        midnightOnly: { type: boolean }
        createdAt: { type: string }
    PostInput:
      type: object
      required: [authorId, content, type]
      properties:
        authorId: { type: integer }
        content: { type: string }
        type: { type: string, enum: [photo, video, text] }
        mediaUrl: { type: string }
        tags: { type: array, items: { type: string } }
        midnightOnly: { type: boolean }
...
    Reel:
      type: object
      required: [id, author, videoUrl, caption, likesCount, viewsCount, createdAt]
      properties:
        id: { type: integer }
        author:
          $ref: "#/components/schemas/User"
        videoUrl: { type: string }
        thumbnailUrl: { type: ["string", "null"] }
        caption: { type: string }
        audioTrack: { type: ["string", "null"] }
        duration: { type: integer }
        likesCount: { type: integer }
        commentsCount: { type: integer }
        viewsCount: { type: integer }
        isLiked: { type: boolean }
        tags: { type: array, items: { type: string } }
        createdAt: { type: string }
        views24h: { type: integer }
    ReelInput:
      type: object
      required: [authorId, videoUrl, caption]
      properties:
        authorId: { type: integer }
        videoUrl: { type: string }
        thumbnailUrl: { type: string }
        caption: { type: string }
        audioTrack: { type: string }
        duration: { type: integer }
        tags: { type: array, items: { type: string } }
...
    Story:
      type: object
      required: [id, author, mediaUrl, mediaType, viewsCount, expiresAt, createdAt]
      properties:
        id: { type: integer }
        author:
          $ref: "#/components/schemas/User"
        mediaUrl: { type: string }
        mediaType: { type: string, enum: [photo, video] }
        caption: { type: ["string", "null"] }
        viewsCount: { type: integer }
        isViewed: { type: boolean }
        expiresAt: { type: string }
        createdAt: { type: string }
    StoryInput:
      type: object
      required: [authorId, mediaUrl, mediaType]
      properties:
        authorId: { type: integer }
        mediaUrl: { type: string }
        mediaType: { type: string, enum: [photo, video] }
        caption: { type: string }
...
    Group:
      type: object
      required: [id, name, description, membersCount, isPrivate, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        avatarUrl: { type: ["string", "null"] }
        coverUrl: { type: ["string", "null"] }
        membersCount: { type: integer }
        postsCount: { type: integer }
        isPrivate: { type: boolean }
        isMember: { type: boolean }
        category: { type: string }
        createdAt: { type: string }
        privacyLevel: { type: string }
        joinType: { type: string }
        groupType: { type: string }
        icon: { type: string }
        themeColor: { type: string }
        maxMembers: { type: integer }
        settings: { type: object }
        creatorId: { type: ["integer", "null"] }
    GroupInput:
      type: object
      required: [name, description]
      properties:
        name: { type: string }
        description: { type: string }
        coverUrl: { type: string }
        avatarUrl: { type: string }
        isPrivate: { type: boolean }
        privacyLevel: { type: string }
        joinType: { type: string }
        category: { type: string }
        groupType: { type: string }
        icon: { type: string }
        themeColor: { type: string }
        maxMembers: { type: integer }
        settings: { type: object }
...
    CreatorPlan:
      type: object
      required: [id, creatorId, name, price, isActive, subscriberCount, createdAt]
      properties:
        id: { type: integer }
        creatorId: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        price: { type: integer }
        perks:
          type: array
          items: { type: string }
        isActive: { type: boolean }
        subscriberCount: { type: integer }
        createdAt: { type: string }
    CreatorPlanInput:
      type: object
      required: [name, price]
      properties:
        name: { type: string }
        description: { type: string }
        price: { type: integer }
        perks:
          type: array
          items: { type: string }
...
    Notification:
      type: object
      required: [id, type, message, isRead, createdAt]
      properties:
        id: { type: integer }
        type: { type: string, enum: [like, comment, follow, mention, group, system] }
        message: { type: string }
        actorName: { type: ["string", "null"] }
        actorAvatar: { type: ["string", "null"] }
        targetId: { type: ["integer", "null"] }
        targetType: { type: ["string", "null"] }
        isRead: { type: boolean }
        createdAt: { type: string }

    # ── AI ────────────────────────────────────────────────────────────────────
...
    AdminUser:
      type: object
      required: [id, username, displayName, email, status, postsCount, followersCount, createdAt]
      properties:
        id: { type: integer }
        username: { type: string }
        displayName: { type: string }
        email: { type: string }
        status: { type: string, enum: [active, suspended] }
        postsCount: { type: integer }
        followersCount: { type: integer }
        isVerified: { type: boolean }
        isAdmin: { type: boolean }
        isPremium: { type: boolean }
        lastSeen: { type: ["string", "null"] }
        createdAt: { type: string }
...
    SearchResults:
      type: object
      properties:
        query: { type: string }
        users:
          type: array
          items: { type: object }
        posts:
          type: array
          items: { type: object }
        reels:
          type: array
          items: { type: object }
        products:
          type: array
          items: { type: object }
...